Fix (quant_tensor): remove duplication (#1204)
